ai/pyproject.toml
[project]
name = "AVALON"
version = "0.1.0"
description = "AVALON AI"
authors = [
    { name = "dig04214", email = "dig04214@naver.com" },
]
requires-python = ">=3.11, <3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "langchain-community>=0.3.25",
    "langchain-openai>=0.3.27",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.8",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.34.3",
    "PyYAML>=6.0",
    "pydantic-settings>=2.9.1",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
